More UserRepository operations. First insert-type SQL operation.

+++ spearfish-db/src/main/java/org/jaffamq/persistence/database/repository/mappings/GroupsByUser.java
@@ -7,7 +7,6 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 import static java.sql.Types.BIGINT;
-import static java.sql.Types.INTEGER;
 
 /**
  * Created by urwisy on 12.01.14.
@@ -16,7 +15,7 @@ public class GroupsByUser extends SelectOperation<Group> {
 
     public GroupsByUser() {
         super("GroupsByUser", "SELECT g.id AS id, g.name AS name FROM security_user_and_group suag" +
-                " INNER JOIN security_group g ON (suag.id_group = g.id) WHERE id_user=?", INTEGER);
+                " INNER JOIN security_group g ON (suag.id_group = g.id) WHERE id_user=?", BIGINT);
     }
 
     @Override
